1.2.1-1.16.1: Changed Dragon Egg recipe.
This is the originally planned recipe. The one from 1.2.0-1.16.1 was
used instead because I'm planning to use a dragon egg to brew dragon
breath and I thought this would allow a loop for infinite dragon breath:
one in the dragon egg and one in the dragon head, opposed to three
obtained by brewing. However, I forgot that I used dragon breath for the
nether star which in turn is used for the elytra, a component of the
dragon egg. So I reversed to the recipe I found more visually appealing
(and also more expensive).